+++ src/roomkit/core/mixins/channel_ops.py
@@ -110,6 +110,7 @@ class ChannelOpsMixin(HelpersMixin):
             channel._realtime = self._realtime
             channel._tool_call_hook = self._build_tool_call_hook(channel.channel_id)
             channel._before_tool_call_hook = self._build_before_tool_call_hook(channel.channel_id)
+            channel._tool_usage_loader = self._build_tool_usage_loader()
             channel._after_response_hook = self._build_after_response_hook(channel.channel_id)
             channel._before_generation_hook = self._build_before_generation_hook(
                 channel.channel_id